fix(b2b-generator): stawka godzinowa ułamkowa (135,5) — koniec z 422
Formularz (input type="number") wysyłał stawkę jako float „135.5", a
B2BRenderRequest.rate_candidate był `Optional[int]` → Pydantic odrzucał
liczbę z częścią ułamkową → 422 „Request failed". Całkowite stawki (150)
przechodziły, ułamkowe zawsze padały.

- schema: rate_candidate → Optional[float] (ścieżka standalone /render;
  /generate zostaje int bo pisze do kolumny Integer Contract).
- format_rate(): wyświetlanie w umowie po polsku — 150 → „150",
  135.5 → „135,50".
- rate_in_words(): dolicza grosze, by słownie zgadzało się z kwotą
  (135,50 → „sto trzydzieści pięć złotych pięćdziesiąt groszy") — wymóg
  spójności umowy.
- FE: defensywna normalizacja przecinek→kropka w stawce.
- testy: grosze (PL/EN), format_rate, regresja render ułamkowej stawki.

+++ backend/app/schemas/b2b_contract_generator.py
@@ -131,7 +131,11 @@ class B2BRenderRequest(BaseModel):
     # Tryb daty rozpoczęcia w §13: "exact" → „z dniem <data>";
     # "not_earlier" → „nie wcześniej niż <data>"; "not_later" → „nie później niż".
     start_date_mode: str = "exact"
-    rate_candidate: Optional[int] = None
+    # Stawka godzinowa — może być ułamkowa (np. 135,5 PLN/h). `float`, NIE `int`:
+    # input type="number" w formularzu wysyła „135.5", a Pydantic `int` odrzuca
+    # liczbę z częścią ułamkową → 422. Formatowanie do postaci „135,50" robi
+    # `format_rate` w `render_context`.
+    rate_candidate: Optional[float] = None
     currency: str = "PLN"
     rate_in_words: Optional[str] = None
     scope_items_override: Optional[list[str]] = None
